refactor: create types.ts file
